requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.26.45
botocore==1.29.45
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.1
cssselect2==0.7.0
decorator==5.1.1
Django==4.1.5
django-better-admin-arrayfield==1.4.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
executing==1.2.0
html5lib==1.1
idna==3.4
inflection==0.5.1
ipython==8.8.0
jedi==0.18.2
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.2.4
lxml==4.9.2
Markdown==3.4.1
matplotlib-inline==0.1.6
oscrypto==1.3.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pypdf==3.7.1
pypng==0.20220715.0
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
s3transfer==0.6.0
sentry-sdk==1.14.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
traitlets==5.8.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.9
